+++ three/public/reset.css
@@ -132,5 +132,5 @@ q:after {
 table {
   border-collapse: collapse;
   border-spacing: 0;
-  /* background-color: #797676; */
+  /* background-color: rgb(215, 63, 63) */
 }
