+++ three/src/lib/objects/ceiling.js
@@ -14,15 +14,17 @@ class Ceiling {
     width = planeWidth,
     length = planeLength,
     color = ceilingColor,
+    rotationY,
   }) {
     const ceilingGeometry = new THREE.PlaneGeometry(width, length);
     const ceilingMaterial = new THREE.MeshBasicMaterial({
       color: color,
-      side: THREE.DoubleSide,
+      side: THREE.BackSide,
     });
     const ceiling = new THREE.Mesh(ceilingGeometry, ceilingMaterial);
     ceiling.name = ceilingName;
     ceiling.rotation.x = Math.PI / 2;
+    ceiling.rotation.y = rotationY;
     // ceiling.rotation.z = Math.PI / 4;
     ceiling.position.set(x, y, z);
